len from int to size_t

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: lpoujade <lpoujade@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/23 21:45:54 by lpoujade          #+#    #+#             */
-/*   Updated: 2016/10/22 17:47:12 by lpoujade         ###   ########.fr       */
+/*   Updated: 2016/11/23 15:46:01 by lpoujade         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -68,7 +68,7 @@ int					ft_strclchr(char const *s, char c);
 char				*ft_strsub(char const *s, unsigned int start, size_t len);
 char				*ft_strrmstr(char *str, char *torm);
 char				*ft_strdup(const char *s1);
-char				*ft_strndup(char const *str, int len);
+char				*ft_strndup(char const *str, size_t len);
 char				*ft_strtrim(char const *s, char *set);
 int					ft_strnequ(char const *s1, char const *s2, size_t n);
 char				*ft_strstr(char const *s1, char const *s2);
